stock-crawler/requirements.txt
akshare>=1.12.0
pandas>=1.5.0
pymysql>=1.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
schedule>=1.2.0
requests>=2.28.0
numpy>=1.24.0
lxml>=4.9.0
beautifulsoup4>=4.11.0
openpyxl>=3.1.0